Room list: using new chat page
